Beginnings of script editing and improvements
Integrated RSyntaxTextArea and added a Haxe syntax highlighting for it,
also added hscript for parsing and interpretation of haxe code. Added a
bunch of UI improvements, including being able to change the base tile
size and resolution. Due to testing of these features, the code is
currently unstable, with some hacks in the client code.