Haxe Compiler Integration
Started working on integrating the haxe compiler. This will allow the
projects to compile to any target from scratch and edit all of the core
and plugin code. I added a subset of the core code to a project folder
which automatically copies to all new projects. There is likely to be
bugs and problems with this, and the project now depends on Haxe, NME
and Neko being installed.